(PC-25410)[API] refactor: add get_payload_from_data to mail backends
