+++ api/src/pcapi/core/mails/backends/testing.py
@@ -19,48 +19,6 @@ class FakeResponse:
         return {"messageId": str(uuid.uuid4())}
 
 
-def get_payload_from_data(
-    recipients: Iterable[str],
-    data: models.TransactionalEmailData | models.TransactionalWithoutTemplateEmailData,
-    bcc_recipients: Iterable[str] = (),
-) -> sendinblue_tasks.SendTransactionalEmailRequest:
-    recipients = list(recipients)
-    bcc_recipients = list(bcc_recipients)
-    reply_to = asdict(data.reply_to)  # equal to sender if reply_to is None
-    scheduled_at = data.scheduled_at
-    if isinstance(data, models.TransactionalEmailData):
-        return sendinblue_tasks.SendTransactionalEmailRequest(
-            attachment=None,
-            bcc_recipients=bcc_recipients,
-            html_content=None,
-            params=data.params,
-            recipients=recipients,
-            reply_to=reply_to,
-            scheduled_at=scheduled_at,
-            sender=asdict(data.template.sender.value),
-            subject=None,
-            tags=data.template.tags,
-            template_id=data.template.id,
-        )
-
-    if isinstance(data, models.TransactionalWithoutTemplateEmailData):
-        return sendinblue_tasks.SendTransactionalEmailRequest(
-            attachment=asdict(data.attachment) if data.attachment else None,
-            bcc_recipients=bcc_recipients,
-            html_content=data.html_content,
-            params=None,
-            recipients=recipients,
-            reply_to=reply_to,
-            scheduled_at=scheduled_at,
-            sender=asdict(data.sender.value),
-            subject=data.subject,
-            tags=None,
-            template_id=None,
-        )
-
-    raise ValueError()
-
-
 class TestingBackend(BaseBackend):
     """A backend that stores email in a global Python list that is
     accessible from tests.
@@ -73,7 +31,7 @@ class TestingBackend(BaseBackend):
         bcc_recipients: Iterable[str] = (),
     ) -> None:
         response = FakeResponse()
-        payload = get_payload_from_data(recipients, data, bcc_recipients)
+        payload, _ = self.get_payload_from_data(recipients, data, bcc_recipients)
         save_message_ids_for_scheduled_emails(response, payload=payload)
 
         sent_data = asdict(data)
